Fix distribution delay and update task names
